change payment rate's hour and rate to double

+++ src/Class/AttendanceReport.cs
@@ -10,16 +10,16 @@ namespace FaceIDAppVBEta.Class
         public int EmployeeNumber { get; set; }
         public DateTime WorkFrom { get; set; }
         public DateTime WorkTo { get; set; }
-        public int RegularHour { get; set; }
-        public int RegularRate { get; set; }
-        public int OvertimeHour1 { get; set; }
-        public int OvertimeRate1 { get; set; }
-        public int OvertimeHour2 { get; set; }
-        public int OvertimeRate2 { get; set; }
-        public int OvertimeHour3 { get; set; }
-        public int OvertimeRate3 { get; set; }
-        public int OvertimeHour4 { get; set; }
-        public int OvertimeRate4 { get; set; }
+        public double RegularHour { get; set; }
+        public double RegularRate { get; set; }
+        public double OvertimeHour1 { get; set; }
+        public double OvertimeRate1 { get; set; }
+        public double OvertimeHour2 { get; set; }
+        public double OvertimeRate2 { get; set; }
+        public double OvertimeHour3 { get; set; }
+        public double OvertimeRate3 { get; set; }
+        public double OvertimeHour4 { get; set; }
+        public double OvertimeRate4 { get; set; }
         public int DayTypeID { get; set; }
         public int PayPeriodID { get; set; }
     }
